Share common ctrl array updates
