+++ grid-community-modules/react/src/reactUi/utils.tsx
@@ -59,3 +59,62 @@ export const agFlushSync = (useFlushSync: boolean, fn: () => void) => {
         fn();
     }
 }
+
+/**
+ * The aim of this function is to maintain references to prev or next values where possible.
+ * If there are not real changes then return the prev value to avoid unnecessary renders.
+ * @param maintainOrder If we want to maintain the order of the elements in the dom in line with the next array
+ * @returns 
+ */
+export function getNextValue<T extends { getInstanceId: () => string }>(prev: T[] | null, next: T[] | null, maintainOrder: boolean): T[] | null {
+
+    if (next == null || prev == null) {
+        return next;
+    }
+
+    // If same array instance nothing to do.
+    // If both empty arrays maintain reference of prev.
+    if (prev === next || (next.length === 0 && prev.length === 0)) {
+        return prev;
+    }
+
+    // If maintaining dom order just return next
+    // If prev is empty just return next immediately as no previous order to maintain
+    // If prev was not empty but next is empty return next immediately
+    if (maintainOrder || (prev.length === 0 && next.length > 0) || (prev.length > 0 && next.length > 0)) {
+        return next;
+    }
+
+    // if dom order not important, we don't want to change the order
+    // of the elements in the dom, as this would break transition styles
+    const oldValues: T[] = [];
+    const newValues: T[] = [];
+    const prevMap: Map<string, T> = new Map();
+
+    for (let i = 0; i < prev.length; i++) {
+        const c = prev[i];
+        prevMap.set(c.getInstanceId(), c)
+    }
+
+    for (let i = 0; i < next.length; i++) {
+        const c = next[i];
+        const instanceId = c.getInstanceId();
+        if (prevMap.has(instanceId)) {
+            oldValues.push(c);
+        } else {
+            newValues.push(c)
+        }
+    }
+
+    // All the same values exist just maybe in a different order so maintain the existing reference
+    if (oldValues.length === prev.length && newValues.length === 0) {
+        return prev;
+    }
+
+    // All new values so avoid spreading the new array to maintain the reference
+    if (oldValues.length === 0 && newValues.length === next.length) {
+        return next;
+    }
+    // Spread as we need to combine the old and new values
+    return [...oldValues, ...newValues];
+}
